LayoutTests/pointerevents/pointer-event-constructors-twist-expected.txt

PASS event.twist is 0
PASS event.twist is 50
PASS successfullyParsed is true

TEST COMPLETE
